+++ src/Application.php
@@ -48,6 +48,7 @@ class Application extends BaseApplication
             new Command\StatsSend(),
             new Command\TorrentClean(),
             new Command\TorrentList(),
+            new Command\TorrentRemove(),
             new Command\TorrentRemoveDuplicates(),
             new Command\WeburgDownload(),
             new Command\WeburgSeriesAdd(),